test-requirements.txt
pytest-ckan
mock
